final/requirements.txt
python>=3.8

numpy>=1.26.4
pandas>=2.1.4
scikit-learn>=1.3.2
scipy>=1.11.4
joblib>=1.3.2
streamlit>=1.31.1
plotly>=5.18.0
matplotlib>=3.8.2
shap>=0.44.0
tensorflow>=2.18.1 ; platform_system != "Windows" or platform_machine != "AMD64"
tensorflow-cpu>=2.18.1 ; platform_system == "Windows" and platform_machine == "AMD64"
pillow>=10.1.0
seaborn>=0.13.2
xgboost>=2.1.2
openai
python-dotenv
